Adds itest for `easel alimanip`.
Fixes two bugs in hidden developer options for alimanip. 1: --xmask
faulted because of a double free. 2: --c-* cluster options faulted
because of writing a free'd MSA.

miniapp-optcheck revised to recognize deliberately undocumented
(developer) options, by checking for high-valued docgroups.
It now checks that undocumented options are indeed undocumented in the
man page, but allows them to be tested in the itest.py script.

esl_getopts.md now documents the convention of using high-valued
docgroups for hidden options.

Adds documentation of the --cindi option to the man page, and adds
some extra explanations for other options.

Added a note to the man page that with no options, alimanip does
nothing and regurgitates the input MSA.

Adds example-rna.sto and example-rfam.sto example files to testsuite:
single and multiple RNA MSAs with #=GC RF annotation, from Rfam seeds.

Made a few arguments to alimanip internal functions `const`, as I was
trying to understand and clarify which args were being modified in the
function call, vs. which were untouched.

Fixes some typos in .c and .man.in.

The hidden --c-* clustering options to `easel alimanip` may not be
working as intended. For example, --cn-id 2 splits example-rna.sto not
into 2 clusters, but to 116 individual sequences. I may be
misunderstanding what the option's supposed to do. Left it as is, and
EPN might have a look sometime.