Fixes a subtle bug recently introduced into esl_sqascii_Parse()
In hmmer3, the hmmpgmd_ga and hmmpgmd_shard_ga tests were failing,
because of a failure in esl_sqascii_Parse().

analyzer. One of those changes subtly broke esl_sqascii_Parse(),
because I misunderstood how the return status of a seebuf() call
needed to be checked ... that eslEOD was not an error, so the original
check for status == eslEFORMAT was correct, and status != eslOK breaks
the code. Rewrote the code to fix it.