Adds easel-compalign-itest.py test script for `easel compalign`.
Makes some tweaks to the help messages for `easel compalign` options,
and to its man page. Adds documentation for --c2dfile to man page.

-p requires that all sequences have PP annotation. There was a check
that msa->pp was non-NULL, but individual PP's could still be missing,
and if so, we were segfaulting. Added a check to fix this, and updated
man page.