Adds itest for `easel ssdraw`.
Adds documentation of --cons, --tabfile to easel ssdraw man page.

Adds `outfile=` optional argument to esl_itest.run(). This specifies
that instead of returning the subprocess.CompletedProcess object, we
slurp the contents of a specified output file and return that as a
string (in UTF-8). This clarifies a bunch of the easel ssdraw tests,
which all look at the output .ps file.

Improves parsing of an optcheck assertion line in miniapp-optcheck.py:
allows comment to be indented, allows trailing whitespace.

Started using pylint. This itest is almost pylint-clean. Had to
struggle a bit to get under pylint's line length restrictions, and
also its restriction on <1000 lines. To be 100% pylint-clean, will
need to systematically change {easel} to {EASEL}, and filenames from
e.g. easel-ssdraw-itest.py to snakecase easel_ssdraw_itest.py. Save
that for the future.